rag_app/requirements.txt
streamlit
langchain
langchain-community
chromadb
transformers
sentence-transformers
torch
nomic
pymupdf
pdf2image
pillow
python-docx
easyocr
numpy
tqdm
openai
ollama
huggingface-hub